test(pack-memory): hoist the shared deterministic embedder fixture into test_support
Three test modules each carried a near-identical FNV-hash embedder fixture. Consolidate them into a
single crate::test_support module and repoint the call sites, so the fixture has one definition.
Test-only, no behavior change.
